+++ hooks/src/lib/index.ts
@@ -1,8 +1,8 @@
-export { useCardCompany } from './hooks/useCardCompany';
-export { useCardCVC } from './hooks/useCardCVC';
-export { useCardExpiryPeriod } from './hooks/useCardExpiryPeriod';
-export { useCardNumber } from './hooks/useCardNumber';
-export { useCardPassword } from './hooks/useCardPassword';
+export { useCardCompany } from './useCardCompany/useCardCompany';
+export { useCardCVC } from './useCardCVC/useCardCVC';
+export { useCardExpiryPeriod } from './useCardExpiryPeriod/useCardExpiryPeriod';
+export { useCardNumber } from './useCardNumber/useCardNumber';
+export { useCardPassword } from './useCardPassword/useCardPassword';
 
 export * from './types/cardTypes';
 export { isOnlyDigits } from './utils/validateNumber';
